fix: addressing review comments
