+++ darts/models/forecasting/forecasting_model.py
@@ -464,157 +464,6 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
             min_future_cov_lag if min_future_cov_lag else 0,
         )
 
-    def _get_historical_forecastable_time_index(
-        self,
-        series: TimeSeries,
-        past_covariates: Optional[TimeSeries] = None,
-        future_covariates: Optional[TimeSeries] = None,
-        is_training: Optional[bool] = False,
-        reduce_to_bounds: bool = False,
-    ) -> Union[
-        pd.DatetimeIndex,
-        pd.RangeIndex,
-        Tuple[int, int],
-        Tuple[pd.Timestamp, pd.Timestamp],
-        None,
-    ]:
-        """
-        Private function that returns the largest time_index representing the subset of each timestamps
-        for which historical forecasts can be made, given the model's properties, the training series
-        and the covariates.
-            - If ``None`` is returned, there is no point where a forecast can be made.
-
-            - If ``is_training=False``, it returns the time_index subset of predictable timestamps.
-
-            - If ``is_training=True``, it returns the time_index subset of trainable timestamps. A trainable
-            timestamp is a timestamp that has a training sample of length at least ``self.training_sample_length``
-             preceding it.
-
-
-        Parameters
-        ----------
-        series
-            A target series.
-        past_covariates
-            Optionally, a past covariates.
-        future_covariates
-            Optionally, a future covariates.
-        is_training
-            Whether the returned time_index should be taking into account the training.
-        reduce_to_bounds
-            Whether to only return the minimum and maximum historical forecastable index
-
-        Returns
-        -------
-        Union[pd.DatetimeIndex, pd.RangeIndex, Tuple[int, int], Tuple[pd.Timestamp, pd.Timestamp], None]
-            The longest time_index that can be used for historical forecasting, either as a range or a tuple.
-
-        Examples
-        --------
-        >>> model = LinearRegressionModel(lags=3, output_chunk_length=2)
-        >>> model.fit(train_series)
-        >>> series = TimeSeries.from_times_and_values(pd.date_range('2000-01-01', '2000-01-10'), np.arange(10))
-        >>> model._get_historical_forecastable_time_index(series=series, is_training=False)
-        DatetimeIndex(['2000-01-04', '2000-01-05', '2000-01-06', '2000-01-07',
-               '2000-01-08', '2000-01-09', '2000-01-10'],
-              dtype='datetime64[ns]', freq='D')
-        >>> model._get_historical_forecastable_time_index(series=series, is_training=True)
-        DatetimeIndex(['2000-01-06', '2000-01-08', '2000-01-09', '2000-01-10'],
-                dtype='datetime64[ns]', freq='D')
-
-        >>> model = NBEATSModel(input_chunk_length=3, output_chunk_length=3)
-        >>> model.fit(train_series, train_past_covariates)
-        >>> series = TimeSeries.from_times_and_values(pd.date_range('2000-10-01', '2000-10-09'), np.arange(8))
-        >>> past_covariates = TimeSeries.from_times_and_values(pd.date_range('2000-10-03', '2000-10-20'),
-        np.arange(18))
-        >>> model._get_historical_forecastable_time_index(series=series, past_covariates=past_covariates,
-        is_training=False)
-        DatetimeIndex(['2000-10-06', '2000-10-07', '2000-10-08', '2000-10-09'], dtype='datetime64[ns]', freq='D')
-        >>> model._get_historical_forecastable_time_index(series=series, past_covariates=past_covariates,
-        is_training=True)
-        DatetimeIndex(['2000-10-09'], dtype='datetime64[ns]', freq='D') # Only one point is trainable, and
-        # corresponds to the first point after we reach a common subset of timestamps of training_sample_length length.
-        """
-
-        (
-            min_target_lag,
-            max_target_lag,
-            min_past_cov_lag,
-            max_past_cov_lag,
-            min_future_cov_lag,
-            max_future_cov_lag,
-        ) = self.extreme_lags
-
-        if min_target_lag is None:
-            min_target_lag = 0
-
-        # longest possible time index for target
-        start = (
-            series.start_time() + (max_target_lag - min_target_lag + 1) * series.freq
-            if is_training
-            else series.start_time() - min_target_lag * series.freq
-        )
-        end = series.end_time() + 1 * series.freq
-
-        intersect_ = (start, end)
-
-        # longest possible time index for past covariates
-        if (min_past_cov_lag is not None) and (past_covariates is not None):
-            start_pc = (
-                past_covariates.start_time()
-                - (min_past_cov_lag - max_target_lag - 1) * past_covariates.freq
-                if is_training
-                else past_covariates.start_time()
-                - min_past_cov_lag * past_covariates.freq
-            )
-            end_pc = (
-                past_covariates.end_time() - max_past_cov_lag * past_covariates.freq
-            )
-            tmp_ = (start_pc, end_pc)
-
-            if intersect_ is not None:
-                intersect_ = (
-                    max([intersect_[0], tmp_[0]]),
-                    min([intersect_[1], tmp_[1]]),
-                )
-            else:
-                intersect_ = tmp_
-
-        # longest possible time index for future covariates
-        if (min_future_cov_lag is not None) and (future_covariates is not None):
-            start_fc = (
-                future_covariates.start_time()
-                - (min_future_cov_lag - max_target_lag - 1) * future_covariates.freq
-                if is_training
-                else future_covariates.start_time()
-                - min_future_cov_lag * future_covariates.freq
-            )
-            end_fc = (
-                future_covariates.end_time()
-                - max_future_cov_lag * future_covariates.freq
-            )
-            tmp_ = (start_fc, end_fc)
-
-            if intersect_ is not None:
-                intersect_ = (
-                    max([intersect_[0], tmp_[0]]),
-                    min([intersect_[1], tmp_[1]]),
-                )
-            else:
-                intersect_ = tmp_
-
-        # end comes before the start
-        if intersect_[1] < intersect_[0]:
-            return None
-
-        # generate an index
-        if not reduce_to_bounds:
-            intersect_ = generate_index(
-                start=intersect_[0], end=intersect_[1], freq=series.freq
-            )
-
-        return intersect_ if len(intersect_) > 0 else None
-
     def _generate_new_dates(
         self, n: int, input_series: Optional[TimeSeries] = None
     ) -> Union[pd.DatetimeIndex, pd.RangeIndex]:
@@ -986,7 +835,6 @@ class ForecastingModel(ABC, metaclass=ModelMeta):
                         idx,
                         past_covariates_,
                         future_covariates_,
-                        bool(retrain),
                     )
                 )
